test: check that compilation against different MPS versions produces the same class files
There were LinkageErrors thrown when running the plugin in a different
MPS version than what was used during compilation.

To ensure that the plugin is actually compatible to all MPS versions it
is compiled twice and then the resulting jar file is checked to be
identical. That check fails if some (Kotlin) API changed that compiles
to different byte code.