fix(mps-legacy-sync-plugin): fix initial synchronization of references in models
Previously, a reference was not synced as PNodeReference if the referenced node not was already seen.

The fix is done for the initial sync in the legacy API.
The new API is not used because it does not support callbacks for progress monitor.
Extending the new API is not reasonable because the plugin is currently rewritten.

The same still exists for references between nodes in different models and in different modules. But it is harder to fix.