feat(mps-sync-plugin): synchronizing read-only modules and models (#154)
* feat(mps-sync-plugin): transform read-only model imports

* feat(mps-sync-plugin): transform read-only module dependencies

* feat(mps-sync-plugin): disable the sync and unbind actions on read-only models and modules

* feat(mps-sync-plugin): transform read-only node references

* fix(mps-sync-plugin-lib): serialize MPSModelImportReference by hand

to avoid a serializer not found exception at runtime

* fix(mps-sync-plugin-lib): move creating read-only model import to a separate method

because we have to do some more checks and steps than just simply creating the MPSModelImportReference

* fix(mps-sync-plugin-lib): encapsulate the MPSNodeReference in a NodeReference

otherwise the reference cannot be serialized

* fix(mps-sync-plugin-lib): setting a reference for a read-only node

* fix(mps-sync-plugin-lib): fix cahing read-only model import during modelix traversal

* fix(mps-sync-plugin-lib): model imports can be added and resolved at any time

Track if the ResolvableModelImport has been resolved (i.e. synced to the cloud) and if so then remove it from the pending this. This way, we can sync model imports directly after they were added to the model.

* refactor(mps-sync-plugin-lib): do not use SNode, SModel, SModule in the cache

Because they may change behind-the-scenes without any warning. As a consequence, they will not be found in the cache as keys, even though they did not change. (MODELIX-787)

* feat(mps-sync-plugin-lib): add SRepository as a new field to the ServiceLocator

so the MPS Repository is always directly accessible and we do not have to go to the MPS Project first

* fix(mps-sync-plugin-lib): create mapping link for the read-only model imports

* fix(mps-sync-plugin-lib): remove outgoing model and module dependency mappings

from the cache, when the corresponding ModelImport, ModuleDependency, LanguageDependency, DevKitDependency is removed

* refactor(mps-sync-plugin-lib): reclassify missed node removal error to warning

* fix(mps-sync-plugin-lib): add null-checks where returned object may be null

* refactor(mps-sync-plugin-lib): inline a variable that is used only once